apps/api/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary,pool]
pgvector
google-generativeai
langgraph
langchain-google-genai
upstash-redis
python-dotenv
pydantic
